+++ basic.py
@@ -1,0 +1,5 @@
+import numpy as np
+import os
+from paper import PaperMatrix, Plan, EagerMatrix
+
+DATA_DIR = "data"
