+++ paper/__init__.py
@@ -1,0 +1,2 @@
+from .core import PaperMatrix
+from .plan import Plan, EagerMatrix
